mode-3-backend/requirements.txt
google-genai
pyaudio
mss
Pillow
pynput
python-dotenv
